Fix delete and sync methods.
Fix for RusDocument<T> to actually sync / delete when appropriate. This was caused by Uuid's not being serialized properly into a document, causing it to not match the _id on the server side.
Updated Cargo.toml to fix license warning on build, and bump version.
Modified the single test I use personally for testing functionality. Sorry it's not good.